Restuctured the source code
