+++ app.js
@@ -1,0 +1,19 @@
+// Run Discord Bot
+require('./bot.js');
+
+// Reserved for later health monitoring
+// Run Express Server (npm install express)
+/*
+const express = require('express')
+const app = express();
+
+app.get('/ping', (req, res) => {
+    console.log('Server was pinged');
+    res.send('Hello World');
+});
+
+const port = process.env.PORT || 8800;
+
+app.listen(port, () => {
+    console.log(`listening on port, ${port}`);
+});*/
