+++ lang/default.js
@@ -1,0 +1,8 @@
+// temp lang system for generic messages
+module.exports = {
+    MESSAGES: {
+        DISCORD: {
+            UNABLE_TO_CONNECT_TO_BACKEND: 'I\'m sorry for the inconvenience but it seems I currently cannot connect to the mother ship.\nPlease try again later.',
+        },
+    },
+};
